Parse available Fedora versions, default to most recent three
There is a REST API which could be parsed to check whether the latest is
released or merely in development, but I'd already written the screen-scraper
before finding that. If a website change breaks this, then the package-version
parsing will probably break at the same time anyway.

The page website's $baseurl contains lines such as
    <img src="/icons/folder.gif" alt="[DIR]"> <a href="11/">11/</a>                     2013-04-25 08:48    -